Extract extensions-location from extensions-impl
For some god awful reason, all of our polyfills were being pulled into
the SW code because of the dependency chains.
